healthcare-agent/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
groq==0.11.0
pdfplumber==0.11.0
python-multipart==0.0.9
python-dotenv==1.0.1